use netip.AddrPort and net.UDPConn (#5)
This change uses the netip.AddrPort type instead of the net.Addr interface type both internally and in the exported package interface. Among other storage and allocation benefits, this greatly simplifies encoding and decoding. In order to better take advantage of package net's new methods to handle netip.AddrPort values, the PacketConn is also hard-coded as a UDP connection.