fix bug acting on messages about deleted IDs (#12)
memberList's update method has had a bug lurking since the initial sketch of the
implementation: the way it used the (increasingly unwieldy) add method could
allow a message about a removed ID to erroneously re-establish that ID in the
membership. With update properly refactored, add only has one call site, and
inlining it eliminates the nonlocality that enabled this behavior.