propagate address values with IDs throughout fsm (#16)
This commit _addresses_ the unnecessary complexity surrounding how address
values are stored and retrieved in the fsm. The ID-to-address relationship is
surjective and persistent, so addresses should be stored alongside other
ID-specific values (incarnation, etc.), which is straightforward as of #15.

Whenever an ID is sent within a struct type (packet, message, Update), the
corresponding address is now almost always sent along with it as well. One
exception to date has been for packet's ping request Target field. This commit
introduces a TargetAddr field, which is necessary in the case that a ping
request recipient is not acquainted with the request target.