emit Updates directly from within fsm (#20)
Significant plumbing can be done away with by enabling the fsm to send Updates
on the driver's updates channel itself rather than return them to the driver
to be sent indirectly. This change does bend the fsm's implicit contract by
permitting it to dabble in concurrent behavior, but only in a way that always
succeeds immediately without blocking (as bufchan.Chan guarantees).