use logarthmic thoroughness coefficient of 2 (#23)
The number of times each node should send each message grows logarithmically
with the network size. The coefficient denoted by λ in the paper describes the
thoroughness of message dissemination. Specifically, if each message is sent λ
log n times, then after λ log n protocol periods, the number of nodes that have
not received the message tends to approach n^-(2λ-2): setting λ > 1
disseminates messages with high confidence.

In a fully configurable implementation of SWIM, this would be one of the
configurable parameters, with the specific value chosen subject to tradeoffs
between thoroughness on the one hand and network traffic and failure detection
latency on the other. For however long this value is going to remain
hard-coded, 2 seems in practice just as effective as the initial draft's
arbitrary value of 3, and more lightweight.