treat memos as implying sender's liveness (#25)
Under the memo extension to the SWIM protocol, a memo implicitly indicates that
its sender is functioning. This is semantically reflected in the fact that the
only message types a node originates regarding itself are alive messages and
new memos. Making this correspondence explicit by including the sender's
incarnation number with each generated memo renders the memo msgType redundant.