exclude memo data from message queue (#26)
This commit fixes a bug by which memo-bearing messages are erroneously
accepted into the message queue for membership status gossip with their memo
data intact. This is an issue because membership status messages should be
small so that several of them can be disseminated in each packet. This commit
causes any memo data present to be stripped from a message bearing novel
membership information before it is upserted into the message queue.