delete entries from wgs on failure
Moving the s.handleFail call to the end of the removal method makes it clearer
that the fsm has finished recording the id's removal before the lock is
released, and therefore that the deletion from wgs happens before any future
entry into processMsg, where s.handleMemo could be called if id was still a
member. It's also consistent with the calls to s.handleJoin and s.handleMemo
occurring at the ends of their respective blocks.