provide network information via handler callbacks (#27)
This change replaces the Updates channel with a trio of methods OnJoin,
OnMemo, and OnFail to register handlers to be called when corresponding
network activity occurs. Their operations are synchronized internally so that
OnJoin's handler will be called before any of the joining node's memos are
delivered via OnMemo's handler, and all memos a node sends while it is a
member of the network will be delivered before OnFail's handler is called for
it. The use of any is optional.